requirements.txt
altgraph==0.17.2
asn1crypto==0.24.0
autopep8==1.5.7
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
cryptography==2.6.1
cx-Freeze==6.8.1
docopt==0.6.2
entrypoints==0.3
gpiozero==1.6.2
idna==2.6
importlib-metadata==4.8.1
keyring==17.1.1
keyrings.alt==3.1.1
mypy==0.910
mypy-extensions==0.4.3
pipreqs==0.4.10
pycairo==1.20.1
pycodestyle==2.7.0
pycrypto==2.6.1
PyGObject==3.30.4
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
python-apt==1.8.4.3
pytz==2021.1
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.5
ssh-import-id==5.7
tabulate==0.8.9
toml==0.10.2
typed-ast==1.4.3
types-pytz==2021.1.2
typing-extensions==3.10.0.2
urllib3==1.24.1
yarg==0.1.9
zipp==3.6.0
